+++ packages/lib/src/utils/constants/currency-decimals.ts
@@ -10,6 +10,7 @@ const CURRENCY_DECIMALS = {
     GHC: 1,
     GNF: 1,
     KMF: 1,
+    ISK: 1,
     PYG: 1,
     RWF: 1,
     UGX: 1,