+++ packages/lib/storybook/stories/cards/Card.stories.tsx
@@ -3,6 +3,8 @@ import { getStoryContextCheckout } from '../../utils/get-story-context-checkout'
 import { CardConfiguration } from '../../../src/components/Card/types';
 import { Card } from '../../../src';
 import { Container } from '../Container';
+import { searchFunctionExample } from '../../../../playground/src/utils';
+import { CardWith3DS2Redirect } from './cardStoryHelpers/CardWith3DS2Redirect';
 
 type CardStory = StoryConfiguration<CardConfiguration>;
 
@@ -22,7 +24,35 @@ export const Default: CardStory = {
     render: createComponent,
     args: {
         componentConfiguration: {
-            _disableClickToPay: true
+            _disableClickToPay: true,
+            autoFocus: true,
+            // brands: ['mc'],
+            // brandsConfiguration: { visa: { icon: 'http://localhost:3000/nocard.svg', name: 'altVisa' } },
+            challengeWindowSize: '02',
+            // configuration: {socialSecurityNumberMode: 'auto'}
+            // data: {
+            //     holderName: 'J. Smith'
+            // },
+            disableIOSArrowKeys: false,
+            // disclaimerMessage,
+            // doBinLookup: false,
+            enableStoreDetails: false,
+            // exposeExpiryDate: true,
+            forceCompat: false,
+            hasHolderName: false,
+            holderNameRequired: false,
+            hideCVC: false,
+            // keypadFix: false,
+            legacyInputMode: false,
+            maskSecurityCode: false,
+            minimumExpiryDate: null, // e.g. '11/24'
+            // name: '', // Affects Dropin only
+            placeholders: {}, // e.g. { holderName: 'B Bob' }
+            positionHolderNameOnTop: false,
+            showBrandIcon: true,
+            showContextualElement: true
+            // showPayButton: false,
+            // styles: { base: { fontWeight: 300 } },
         }
     }
 };
@@ -34,6 +64,7 @@ export const WithAVS: CardStory = {
             _disableClickToPay: true,
             billingAddressRequired: true,
             billingAddressAllowedCountries: ['US', 'CA', 'GB'],
+            // billingAddressRequiredFields: ['postalCode', 'country'],
             data: {
                 billingAddress: {
                     street: 'Virginia Street',
@@ -55,6 +86,17 @@ export const WithPartialAVS: CardStory = {
             _disableClickToPay: true,
             billingAddressRequired: true,
             billingAddressMode: 'partial'
+        }
+    }
+};
+
+export const WithAVSAddressLookup: CardStory = {
+    render: createComponent,
+    args: {
+        componentConfiguration: {
+            _disableClickToPay: true,
+            billingAddressRequired: true,
+            onAddressLookup: searchFunctionExample
         }
     }
 };
@@ -82,6 +124,7 @@ export const KCP: CardStory = {
     render: createComponent,
     args: {
         componentConfiguration: {
+            ...{ brands: ['mc', 'visa', 'amex', 'bcmc', 'maestro', 'korean_local_card'] },
             _disableClickToPay: true,
             // Set koreanAuthenticationRequired AND countryCode so KCP fields show at start
             // Just set koreanAuthenticationRequired if KCP fields should only show if korean_local_card entered
@@ -102,6 +145,18 @@ export const WithClickToPay: CardStory = {
                 merchantDisplayName: 'Adyen Merchant Name'
             }
         }
+    }
+};
+
+export const CardWith_3DS2_Redirect: CardStory = {
+    render: args => {
+        return <CardWith3DS2Redirect contextArgs={args} />;
+    },
+    args: {
+        componentConfiguration: {
+            _disableClickToPay: true
+        },
+        useSessions: false
     }
 };