+++ packages/lib/storybook/stories/issuer-lists/OnlineBankingPL.stories.tsx
@@ -2,20 +2,20 @@ import { MetaConfiguration, PaymentMethodStoryProps, StoryConfiguration } from '
 import { getStoryContextCheckout } from '../../utils/get-story-context-checkout';
 import { Container } from '../Container';
 import { IssuerListConfiguration } from '../../../src/components/helpers/IssuerListContainer/types';
-import { Dotpay } from '../../../src';
+import { OnlineBankingPL } from '../../../src';
 
-type DotpayStory = StoryConfiguration<IssuerListConfiguration>;
+type OnlineBankingPLStory = StoryConfiguration<IssuerListConfiguration>;
 
 const meta: MetaConfiguration<IssuerListConfiguration> = {
-    title: 'IssuerList/Dotpay'
+    title: 'IssuerList/OnlineBankingPL'
 };
 
-export const Default: DotpayStory = {
+export const Default: OnlineBankingPLStory = {
     render: (args: PaymentMethodStoryProps<IssuerListConfiguration>, context) => {
         const { componentConfiguration } = args;
         const checkout = getStoryContextCheckout(context);
-        const dotpay = new Dotpay(checkout, componentConfiguration);
-        return <Container element={dotpay} />;
+        const onlineBankingPL = new OnlineBankingPL(checkout, componentConfiguration);
+        return <Container element={onlineBankingPL} />;
     },
     args: {
         countryCode: 'PL'